Normalizes height of button to match input when zoomed
Today the button has a height of 17 plus two paddings of 14, total 45.
The input has 19 plus two paddings of 12 and a border of 1, total 45.
It’s ok, but it breaks when you zoom it because it looks like the
browser re-scales the internal height and the padding (and the border!)
differently.